Struct fps::Fps

pub struct Fps<const P: u64> { /* private fields */ }

The formal power series (FPS). This object knows a FPS $f$ modulo $x^d$ where $d$ is precision. The precision can be infinite.

Implementation

This is a pair of a vector of coeffs and precision. The infinite precision is represented by usize::MAX.

Invariants

coeffs.last() != Some(&Fp::new(0))

coeffs.len() <= precision

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::new(vec![fp!(1), fp!(2)], 4);
let g = f.inv();
let expected = Fps::new(vec![fp!(1), fp!(-2), fp!(4), fp!(-8)], 4);
assert_eq!(g, expected);

Implementations

impl<const P: u64> Fps<P>

pub fn new(coeffs: Vec<Fp<P>>, precision: usize) -> Self

Creates a new FPS with the given coefficients and precision. # Panics Panics if the last coefficient is zero.

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::new(vec![fp!(1), fp!(2)], 4);
assert_eq!(f.precision(), 4);
assert_eq!(f.into_coeffs(), vec![fp!(1), fp!(2)]);

pub fn polynominal(coeffs: Vec<Fp<P>>) -> Self

Creates a new FPS the given coefficients and infinite precision.

Panics

Panics if the last coefficient is zero.

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::polynominal(vec![fp!(1), fp!(2)]);
assert_eq!(f.precision(), usize::MAX);
assert_eq!(f.into_coeffs(), vec![fp!(1), fp!(2)]);

pub fn at(&self, i: usize) -> Option<Fp<P>>

pub fn into_coeffs(self) -> Vec<Fp<P>>

Returns the vector of coefficients.

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::new(vec![fp!(1), fp!(2)], 4);
assert_eq!(f.into_coeffs(), vec![fp!(1), fp!(2)]);

pub fn precision(&self) -> usize

Returns the precision $d$.

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::new(vec![fp!(1), fp!(2)], 4);
assert_eq!(f.precision(), 4);

pub fn set_precision(&mut self, precision: usize)

Changes the precision. If the new precision is smaller than the current one, the coefficients are truncated.

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let mut f = Fps::new(vec![fp!(1), fp!(2)], 4);
f.set_precision(1);
assert_eq!(f.precision(), 1);
assert_eq!(f.into_coeffs(), vec![fp!(1)]);
let mut f = Fps::new(vec![fp!(1), fp!(2)], 4);
f.set_precision(10);
assert_eq!(f.precision(), 10);
assert_eq!(f.into_coeffs(), vec![fp!(1), fp!(2)]);

pub fn copied_with(&self, precision: usize) -> Self

Returns the truncated FPS of self.

Complexity

$O(d)$

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::new(vec![fp!(1), fp!(2)], 4);
let g = f.copied_with(1);
assert_eq!(g.precision(), 1);
assert_eq!(g.into_coeffs(), vec![fp!(1)]);

pub fn coeffs(&self) -> impl '_ + Iterator<Item = Fp<P>>

Returns the iterator of coefficients.

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::new(vec![fp!(1), fp!(2)], 4);
assert_eq!(f.coeffs().collect::<Vec<_>>(), vec![fp!(1), fp!(2)]);

pub fn coeffs_mut(&mut self) -> impl Iterator<Item = &mut Fp<P>>

Returns the mutable iterator of coefficients.

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let mut f = Fps::new(vec![fp!(1), fp!(2)], 4);
for a in f.coeffs_mut() {
   *a += fp!(1);
}
assert_eq!(f.into_coeffs(), vec![fp!(2), fp!(3)]);

impl<const P: u64> Fps<P>where (): PrimitiveRoot<P>,

pub fn inv(&self) -> Self

Inverse FPS of self.

Requirements

$f_0 \ne 0$

Returns

$f^{-1} \mod x^d$

Complexity

$O(d \log d)$

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::new(vec![fp!(1), fp!(2)], 4);
let g = f.inv();
let expected = Fps::new(vec![fp!(1), fp!(-2), fp!(4), fp!(-8)], 4);
assert_eq!(g, expected);

pub fn sqrt(&self) -> Self

Square root of self.

Requirements

$f_0 = 1$

Returns

$f^{1/2} \mod x^d$

Complexity

$O(d \log d)$

Examples

use fp2::fp;
type Fps = fp2::Fps<998244353>;
let f = Fps::new(vec![fp!(1), fp!(2), fp!(3), fp!(4), fp!(5)], 5);
let g = f.sqrt();
let expected = Fps::new(vec![fp!(1); 5], 5);
assert_eq!(g, expected);