Crate fps

Arithmetic of formal power series.

Note on complexity

We only consider cases where the precision is a power of 2. If the precision is not a power of 2, the complexity is twice as bad.

We may assume the complexity $\mathcal{M}(d)$ of multiplication of two polynomials of degree $d$ and the complexity $\mathcal{F}(d)$ of FFT/IFFT of length $d$ satisfies the following property:

$$ \begin{aligned} \mathcal{M}(d) &= \Theta(d\log d) , \mathcal{F}(d) = \Theta(d\log d), \\ \mathcal{M}(d) &= 3 \mathcal{F}(2d) + O(d) \ \left( = 6 \mathcal{F}(d) + O(d) \right). \end{aligned} $$

If each step in Newton’s method is performed in $a\mathcal{M}(d)$ time (where $d$ is the resulting precision of each iteration), then the total complexity is $2a\mathcal{F}(d) + O(d)$.

We omit $O(d)$ terms when it is not important.

$O(d)$ is omitted here.

Name	Expression	Complexity

fps_inv	$f^{-1} \mod x^d$	$2\mathcal{M}(d)$

fps_sqrt	$f^{1/2} \mod x^d$	$6\mathcal{M}(d)$

Functions

fps_inv

Inverse FPS of f.

fps_sqrt

Square root FPS of f.